+++ internal/app/handlers/constants.go
@@ -5,5 +5,5 @@ const (
 	OriginalUserIDKey   = "original_user_id"
 	UserRoleKey         = "user_role"
 	OriginalUserRoleKey = "original_user_role"
-	FlashKey            = "flash"
+	FlashCookiePrefix   = "flash"
 )