+++ internal/backends/interfaces.go
@@ -11,31 +11,30 @@ import (
 )
 
 type Services struct {
-	ORCIDSandbox               bool
-	ORCIDClient                *orcid.MemberClient
-	Repository                 Repository
-	FileStore                  FileStore
-	DatasetSearchService       DatasetSearchService
-	PublicationSearchService   PublicationSearchService
-	OrganizationService        OrganizationService
-	PersonService              PersonService
-	ProjectService             ProjectService
-	UserService                UserService
-	OrganizationSearchService  OrganizationSearchService
-	PersonSearchService        PersonSearchService
-	ProjectSearchService       ProjectSearchService
-	UserSearchService          UserSearchService
-	LicenseSearchService       LicenseSearchService
-	MediaTypeSearchService     MediaTypeSearchService
-	PublicationSources         map[string]PublicationGetter
-	DatasetSources             map[string]DatasetGetter
-	PublicationEncoders        map[string]PublicationEncoder
-	PublicationDecoders        map[string]PublicationDecoderFactory
-	PublicationListExporters   map[string]PublicationListExporterFactory
-	PublicationSearcherService PublicationSearcherService
-	DatasetListExporters       map[string]DatasetListExporterFactory
-	DatasetSearcherService     DatasetSearcherService
-	HandleService              HandleService
+	ORCIDSandbox              bool
+	ORCIDClient               *orcid.MemberClient
+	Repository                Repository
+	FileStore                 FileStore
+	SearchService             SearchService
+	DatasetSearchIndex        DatasetIndex
+	PublicationSearchIndex    PublicationIndex
+	OrganizationService       OrganizationService
+	PersonService             PersonService
+	ProjectService            ProjectService
+	UserService               UserService
+	OrganizationSearchService OrganizationSearchService
+	PersonSearchService       PersonSearchService
+	ProjectSearchService      ProjectSearchService
+	UserSearchService         UserSearchService
+	LicenseSearchService      LicenseSearchService
+	MediaTypeSearchService    MediaTypeSearchService
+	PublicationSources        map[string]PublicationGetter
+	DatasetSources            map[string]DatasetGetter
+	PublicationEncoders       map[string]PublicationEncoder
+	PublicationDecoders       map[string]PublicationDecoderFactory
+	PublicationListExporters  map[string]PublicationListExporterFactory
+	DatasetListExporters      map[string]DatasetListExporterFactory
+	HandleService             HandleService
 	// Tasks                      *tasks.Hub
 }
 
@@ -56,7 +55,6 @@ type PublicationGetter interface {
 }
 
 type Repository interface {
-	Transaction(context.Context, func(Repository) error) error
 	GetPublication(string) (*models.Publication, error)
 	GetPublications([]string) ([]*models.Publication, error)
 	SavePublication(*models.Publication, *models.User) error
@@ -119,38 +117,135 @@ type BulkIndexer[T any] interface {
 	Close(context.Context) error
 }
 
-type DatasetSearchService interface {
+type DatasetIDIndex interface {
+	Search(*models.SearchArgs) (*models.SearchHits, error)
+	Each(searchArgs *models.SearchArgs, maxSize int, cb func(string)) error
+	Delete(id string) error
+	DeleteAll() error
+	WithScope(string, ...string) DatasetIDIndex
+}
+
+type DatasetIndex interface {
 	Search(*models.SearchArgs) (*models.DatasetHits, error)
-	Index(*models.Dataset) error
+	Each(searchArgs *models.SearchArgs, maxSize int, cb func(*models.Dataset)) error
 	Delete(id string) error
 	DeleteAll() error
-	WithScope(string, ...string) DatasetSearchService
-	NewBulkIndexer(BulkIndexerConfig) (BulkIndexer[*models.Dataset], error)
-	NewIndexSwitcher(BulkIndexerConfig) (IndexSwitcher[*models.Dataset], error)
+	WithScope(string, ...string) DatasetIndex
 }
 
-type PublicationSearchService interface {
+type datasetIndex struct {
+	DatasetIDIndex
+	repo Repository
+}
+
+func NewDatasetIndex(di DatasetIDIndex, r Repository) DatasetIndex {
+	return &datasetIndex{
+		DatasetIDIndex: di,
+		repo:           r,
+	}
+}
+
+func (ds *datasetIndex) Search(args *models.SearchArgs) (*models.DatasetHits, error) {
+	h, err := ds.DatasetIDIndex.Search(args)
+	if err != nil {
+		return nil, err
+	}
+
+	datasets, err := ds.repo.GetDatasets(h.Hits)
+	if err != nil {
+		return nil, err
+	}
+
+	return &models.DatasetHits{
+		Pagination: h.Pagination,
+		Hits:       datasets,
+		Facets:     h.Facets,
+	}, nil
+}
+
+func (ds *datasetIndex) Each(searchArgs *models.SearchArgs, maxSize int, cb func(*models.Dataset)) error {
+	return ds.DatasetIDIndex.Each(searchArgs, maxSize, func(id string) {
+		// TODO handle error
+		dataset, _ := ds.repo.GetDataset(id)
+		cb(dataset)
+	})
+}
+
+func (ds *datasetIndex) WithScope(field string, terms ...string) DatasetIndex {
+	return &datasetIndex{
+		DatasetIDIndex: ds.DatasetIDIndex.WithScope(field, terms...),
+		repo:           ds.repo,
+	}
+}
+
+type PublicationIDIndex interface {
+	Search(*models.SearchArgs) (*models.SearchHits, error)
+	Each(searchArgs *models.SearchArgs, maxSize int, cb func(string)) error
+	Delete(id string) error
+	DeleteAll() error
+	WithScope(string, ...string) PublicationIDIndex
+}
+
+type PublicationIndex interface {
 	Search(*models.SearchArgs) (*models.PublicationHits, error)
-	Index(*models.Publication) error
+	Each(searchArgs *models.SearchArgs, maxSize int, cb func(*models.Publication)) error
 	Delete(id string) error
 	DeleteAll() error
-	WithScope(string, ...string) PublicationSearchService
-	NewBulkIndexer(BulkIndexerConfig) (BulkIndexer[*models.Publication], error)
-	NewIndexSwitcher(BulkIndexerConfig) (IndexSwitcher[*models.Publication], error)
+	WithScope(string, ...string) PublicationIndex
 }
 
-type PublicationSearcherService interface {
-	GetMaxSize() int
-	SetMaxSize(int)
-	WithScope(string, ...string) PublicationSearcherService
-	Searcher(*models.SearchArgs, func(*models.Publication)) error
+type publicationIndex struct {
+	PublicationIDIndex
+	repo Repository
 }
 
-type DatasetSearcherService interface {
-	GetMaxSize() int
-	SetMaxSize(int)
-	WithScope(string, ...string) DatasetSearcherService
-	Searcher(*models.SearchArgs, func(*models.Dataset)) error
+func NewPublicationIndex(di PublicationIDIndex, r Repository) PublicationIndex {
+	return &publicationIndex{
+		PublicationIDIndex: di,
+		repo:               r,
+	}
+}
+
+func (pi *publicationIndex) Search(args *models.SearchArgs) (*models.PublicationHits, error) {
+	h, err := pi.PublicationIDIndex.Search(args)
+	if err != nil {
+		return nil, err
+	}
+
+	publications, err := pi.repo.GetPublications(h.Hits)
+	if err != nil {
+		return nil, err
+	}
+
+	return &models.PublicationHits{
+		Pagination: h.Pagination,
+		Hits:       publications,
+		Facets:     h.Facets,
+	}, nil
+}
+
+func (pi *publicationIndex) Each(searchArgs *models.SearchArgs, maxSize int, cb func(*models.Publication)) error {
+	return pi.PublicationIDIndex.Each(searchArgs, maxSize, func(id string) {
+		// TODO handle error
+		publication, _ := pi.repo.GetPublication(id)
+		cb(publication)
+	})
+}
+
+func (pi *publicationIndex) WithScope(field string, terms ...string) PublicationIndex {
+	return &publicationIndex{
+		PublicationIDIndex: pi.PublicationIDIndex.WithScope(field, terms...),
+		repo:               pi.repo,
+	}
+}
+
+type SearchService interface {
+	NewDatasetIndex(Repository) DatasetIndex
+	NewDatasetBulkIndexer(BulkIndexerConfig) (BulkIndexer[*models.Dataset], error)
+	NewDatasetIndexSwitcher(BulkIndexerConfig) (IndexSwitcher[*models.Dataset], error)
+	NewPublicationIndex(Repository) PublicationIndex
+	NewPublicationBulkIndexer(BulkIndexerConfig) (BulkIndexer[*models.Publication], error)
+	NewPublicationIndexSwitcher(BulkIndexerConfig) (IndexSwitcher[*models.Publication], error)
 }
 
 type OrganizationService interface {
@@ -159,7 +254,6 @@ type OrganizationService interface {
 
 type PersonService interface {
 	GetPerson(string) (*models.Person, error)
-	GetPersons([]string) ([]*models.Person, error)
 }
 
 type ProjectService interface {
@@ -176,7 +270,7 @@ type OrganizationSearchService interface {
 }
 
 type PersonSearchService interface {
-	SuggestPeople(string) ([]models.Person, error)
+	SuggestPeople(string) ([]*models.Person, error)
 }
 
 type ProjectSearchService interface {
@@ -184,7 +278,7 @@ type ProjectSearchService interface {
 }
 
 type UserSearchService interface {
-	SuggestUsers(string) ([]models.Person, error)
+	SuggestUsers(string) ([]*models.Person, error)
 }
 
 type LicenseSearchService interface {
@@ -219,20 +313,100 @@ type HandleService interface {
 
 var ErrNotFound = errors.New("record not found")
 
-type RepositoryFilter struct {
-	Field string
-	Op    string
-	Value string
-}
-
-type RepositoryQueryArgs struct {
-	Limit   int
-	Offset  int
-	Order   string
-	Filters []*RepositoryFilter
-}
-
 type Mutation struct {
 	Op   string
 	Args []string
+}
+
+type PersonWithOrganizationsService struct {
+	PersonService       PersonService
+	OrganizationService OrganizationService
+}
+
+func (s *PersonWithOrganizationsService) GetPerson(id string) (*models.Person, error) {
+	p, err := s.PersonService.GetPerson(id)
+	if err != nil {
+		return nil, err
+	}
+	for _, a := range p.Affiliations {
+		o, err := s.OrganizationService.GetOrganization(a.OrganizationID)
+		if err == ErrNotFound {
+			a.Organization = NewDummyOrganization(a.OrganizationID)
+		} else if err != nil {
+			return nil, err
+		} else {
+			a.Organization = o
+		}
+	}
+	return p, nil
+}
+
+type UserWithOrganizationsService struct {
+	UserService         UserService
+	OrganizationService OrganizationService
+}
+
+func (s *UserWithOrganizationsService) GetUser(id string) (*models.User, error) {
+	u, err := s.UserService.GetUser(id)
+	if err != nil {
+		return nil, err
+	}
+	for _, a := range u.Affiliations {
+		o, err := s.OrganizationService.GetOrganization(a.OrganizationID)
+		if err == ErrNotFound {
+			a.Organization = NewDummyOrganization(a.OrganizationID)
+		} else if err != nil {
+			return nil, err
+		} else {
+			a.Organization = o
+		}
+	}
+	return u, nil
+}
+
+func (s *UserWithOrganizationsService) GetUserByUsername(username string) (*models.User, error) {
+	u, err := s.UserService.GetUserByUsername(username)
+	if err != nil {
+		return nil, err
+	}
+	for _, a := range u.Affiliations {
+		o, err := s.OrganizationService.GetOrganization(a.OrganizationID)
+		if err == ErrNotFound {
+			a.Organization = NewDummyOrganization(a.OrganizationID)
+		} else if err != nil {
+			return nil, err
+		} else {
+			a.Organization = o
+		}
+	}
+	return u, nil
+}
+
+// TODO remove this when we always have an uptodate db of organizations
+func NewDummyOrganization(id string) *models.Organization {
+	return &models.Organization{
+		ID:   id,
+		Name: id,
+		Tree: []struct {
+			ID string `json:"id,omitempty"`
+		}{
+			{ID: id},
+		},
+	}
+}
+
+func NewDummyPerson(id string) *models.Person {
+	return &models.Person{
+		ID:        id,
+		FullName:  "[missing]",
+		FirstName: "[missing]",
+		LastName:  "[missing]",
+	}
+}
+
+func NewDummyProject(id string) *models.Project {
+	return &models.Project{
+		ID:    id,
+		Title: "[missing]",
+	}
 }