+++ internal/models/organization.go
@@ -1,11 +1,14 @@
 package models
 
-type OrganizationRef struct {
-	ID string `json:"id,omitempty"`
+type Organization struct {
+	ID   string `json:"id,omitempty"`
+	Name string `json:"name,omitempty"`
+	Tree []struct {
+		ID string `json:"id,omitempty"`
+	} `json:"tree,omitempty"`
 }
 
-type Organization struct {
-	ID   string            `json:"id,omitempty"`
-	Name string            `json:"name,omitempty"`
-	Tree []OrganizationRef `json:"tree,omitempty"`
+type RelatedOrganization struct {
+	OrganizationID string        `json:"organization_id,omitempty"`
+	Organization   *Organization `json:"-"`
 }